module09-kafka.programming.in.python/requirements.txt
pymongo
kafka-python
websockets